♻️🏢 Rename partner folder
