+++ apps/admin/src/show/partners/category.tsx
